+++ tophat@fflewddur.github.io/lib/fs.js
@@ -104,6 +104,13 @@ var FileSystemMonitor = GObject.registerClass({
             GObject.ParamFlags.READWRITE,
             ''
         ),
+        'monitor-mode': GObject.ParamSpec.string(
+            'monitor-mode',
+            'Monitor mode',
+            'Monitor filesystem usage, disk activity, or both',
+            GObject.ParamFlags.READWRITE,
+            ''
+        ),
     },
 }, class TopHatFileSystemMonitor extends Monitor.TopHatMonitor {
     _init(configHandler) {
@@ -120,6 +127,20 @@ var FileSystemMonitor = GObject.registerClass({
         });
         this.add_child(this.usage);
 
+        let vbox = new St.BoxLayout({vertical: true});
+        vbox.connect('notify::vertical', obj => {
+            obj.vertical = true;
+        });
+        this.add_child(vbox);
+        this.activityBox = vbox;
+
+        let valueRead = new St.Label({text: '0', style_class: 'tophat-meter-value-net', y_expand: true, y_align: Clutter.ActorAlign.END});
+        vbox.add_child(valueRead);
+        this.valueRead = valueRead;
+        let valueWrite = new St.Label({text: '0', style_class: 'tophat-meter-value-net', y_expand: true, y_align: Clutter.ActorAlign.START});
+        vbox.add_child(valueWrite);
+        this.valueWrite = valueWrite;
+
         this.timePrev = GLib.get_monotonic_time();
         this.diskActivityPrev = new DiskActivity();
         this.history = new Array(0);
@@ -135,6 +156,7 @@ var FileSystemMonitor = GObject.registerClass({
         configHandler.settings.bind('show-animations', this, 'show-animation', Gio.SettingsBindFlags.GET);
         configHandler.settings.bind('disk-display', this, 'visualization', Gio.SettingsBindFlags.GET);
         configHandler.settings.bind('mount-to-monitor', this, 'mount', Gio.SettingsBindFlags.GET);
+        configHandler.settings.bind('disk-monitor-mode', this, 'monitor-mode', Gio.SettingsBindFlags.GET);
 
         let id = this.connect('notify::visible', () => {
             if (this.visible) {
@@ -147,6 +169,10 @@ var FileSystemMonitor = GObject.registerClass({
         id = this.connect('notify::refresh-rate', () => {
             this._stopTimers();
             this._startTimers();
+        });
+        this._signals.push(id);
+        id = this.connect('notify::monitor-mode', () => {
+            this.updateVisualization();
         });
         this._signals.push(id);
 
@@ -165,6 +191,36 @@ var FileSystemMonitor = GObject.registerClass({
         }
         this._mount = value;
         this.notify('mount');
+    }
+
+    get monitor_mode() {
+        return this._monitor_mode;
+    }
+
+    set monitor_mode(value) {
+        if (this._monitor_mode === value) {
+            return;
+        }
+        this._monitor_mode = value;
+        this.notify('monitor-mode');
+    }
+
+    updateVisualization() {
+        super.updateVisualization();
+
+        switch (this.monitor_mode) {
+        case 'storage':
+            this.activityBox.visible = false;
+            break;
+        case 'activity':
+            this.activityBox.visible = true;
+            this.usage.visible = false;
+            this.meter.visible = false;
+            break;
+        case 'both':
+            this.activityBox.visible = true;
+            break;
+        }
     }
 
     _startTimers() {
@@ -218,20 +274,35 @@ var FileSystemMonitor = GObject.registerClass({
 
         label = new St.Label({text: _('Reading:'), style_class: 'menu-label'});
         this.addMenuRow(label, 0, 2, 1);
-        label = new St.Label({text: _('n/a'), style_class: 'menu-value'});
+        label = new St.Label({text: _('n/a'), style_class: 'menu-value menu-section-end'});
         this.addMenuRow(label, 2, 1, 1);
         this.menuDiskReads = label;
 
-        this.historyChart = new St.DrawingArea({style_class: 'chart'});
+        // Create a grid layout for the history chart
+        let grid = new St.Widget({
+            layout_manager: new Clutter.GridLayout({orientation: Clutter.Orientation.VERTICAL}),
+        });
+        this.historyGrid = grid.layout_manager;
+        this.addMenuRow(grid, 0, 3, 1);
+
+        this.historyChart = new St.DrawingArea({style_class: 'chart', x_expand: true});
         this.historyChart.connect('repaint', () => this._repaintHistory());
-        this.addMenuRow(this.historyChart, 0, 3, 1);
+        this.historyGrid.attach(this.historyChart, 0, 0, 2, 3);
+
+        label = new St.Label({text: _('Write'), y_align: Clutter.ActorAlign.START, style_class: 'chart-label'});
+        this.historyGrid.attach(label, 2, 0, 1, 1);
+        label = new St.Label({text: '100%', y_align: Clutter.ActorAlign.CENTER, style_class: 'chart-label'});
+        this.historyGrid.attach(label, 2, 1, 1, 1);
+        this.historyMaxVal = label;
+        label = new St.Label({text: _('Read'), y_align: Clutter.ActorAlign.END, style_class: 'chart-label'});
+        this.historyGrid.attach(label, 2, 2, 1, 1);
 
         let limitInMins = Config.HISTORY_MAX_SIZE / 60;
         let startLabel = ngettext('%d min ago', '%d mins ago', limitInMins).format(limitInMins);
         label = new St.Label({text: startLabel, style_class: 'chart-label-then'});
-        this.addMenuRow(label, 0, 2, 1);
+        this.historyGrid.attach(label, 0, 3, 1, 1);
         label = new St.Label({text: _('now'), style_class: 'chart-label-now'});
-        this.addMenuRow(label, 2, 1, 1);
+        this.historyGrid.attach(label, 1, 3, 1, 1);
 
         label = new St.Label({text: _('Top processes'), style_class: 'menu-header'});
         this.addMenuRow(label, 0, 3, 1);
@@ -260,7 +331,7 @@ var FileSystemMonitor = GObject.registerClass({
         label = new St.Label({text: _('Filesystem usage'), style_class: 'menu-header'});
         this.addMenuRow(label, 0, 3, 1);
 
-        let grid = new St.Widget({
+        grid = new St.Widget({
             layout_manager: new Clutter.GridLayout({orientation: Clutter.Orientation.VERTICAL}),
         });
         this.menuFSDetails = grid.layout_manager;
@@ -402,6 +473,8 @@ var FileSystemMonitor = GObject.registerClass({
         let diskWrite = Shared.bytesToHumanString(Math.round(write));
         this.menuDiskReads.text = `${diskRead}/s`;
         this.menuDiskWrites.text = `${diskWrite}/s`;
+        this.valueRead.text = `${diskRead}/s`;
+        this.valueWrite.text = `${diskWrite}/s`;
 
         while (this.history.length >= this.historyLimit) {
             this.history.shift();
@@ -420,9 +493,10 @@ var FileSystemMonitor = GObject.registerClass({
         let pointSpacing = width / (this.historyLimit - 1);
         let xStart = (this.historyLimit - this.history.length) * pointSpacing;
         let ctx = this.historyChart.get_context();
-        var fg, bg;
+        var fg, bg, gc;
         [, fg] = Clutter.Color.from_string(this.meter_fg_color);
         [, bg] = Clutter.Color.from_string(Config.METER_BG_COLOR);
+        [, gc] = Clutter.Color.from_string(Config.METER_GRID_COLOR);
 
         // Use a small value to avoid max == 0
         let max = 0.001;
@@ -434,10 +508,16 @@ var FileSystemMonitor = GObject.registerClass({
                 max = da.write;
             }
         }
+        max = Shared.roundMax(max);
+        this.historyMaxVal.text = `${Shared.bytesToHumanString(max, 'bytes', true)}/s`;
         max *= 2; // leave room for both upload and download speeds on the same chart
 
         Clutter.cairo_set_source_color(ctx, bg);
         ctx.rectangle(0, 0, width, height);
+        ctx.fill();
+
+        Clutter.cairo_set_source_color(ctx, gc);
+        ctx.rectangle(0, height / 2, width, 1);
         ctx.fill();
 
         Clutter.cairo_set_source_color(ctx, fg);