+++ RELEASES.md
@@ -3,6 +3,12 @@
 All notable changes to [TopHat] are listed in this file. The format is loosely
 based on [Keep a Changelog].
 
+## TopHat 11 - June 10, 2023
+
+- Added option to show disk activity instead of (or in addition to) available storage
+- Added y-axis scales to the activity over time charts
+- Added system uptime to the CPU monitor menu
+
 ## TopHat 10 - May 29, 2023
 
 - Added option to adjust refresh speed